# oas3_allyoucanorder.yaml
openapi: 3.0.0
info:
  version: '1.0'
  title: "AllYouCanOrder"
  description: API for managing restaurant orders.
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Localhost
paths:
...
  /piattosRisto:
    get:
      description: >-
          Gets the list of dishes on the menu of a specific restaurant. 
      summary: View all dishes
      parameters:
        - name: x-access-token
          in: header
          description: an authorization header
          required: true
          schema:
            type: string 
          example: token.esempio.prova
      responses:
        '200':
          description: 'Collection of dishes'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Piatto'
              example:
                  - nome: 'Gamberi fritti'
                    prezzo: 2.49
                    descrizione: '5 pezzi. Ingredienti: gamberi'
                    foto: 'www.gamberifritti.images.com'
                    stato: 'consegnato'
                  - nome: 'California Roll'
                    prezzo: 2.79
                    descrizione: '5 pezzi. Ingredienti: riso, salmone crudo, philadelphia'
                    foto: 'www.gamberifritti.images.com'
                    stato: 'consegnato'
        '404':
          description: 'Error message: We could not find the dishes you were looking for'
  /piattosRisto/aggiungiPiatto:
    post:
      description: >-
        Creates a new dish in the system.
      summary: Add a new dish in a restaurant's menu
      parameters:
        - name: x-access-token
          in: header
          description: an authorization header
          required: true
          schema:
            type: string 
          example: token.esempio.prova
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - nome
                - prezzo
              properties:
                nome:
                  type: string
                  description: 'Name of the dish'
                prezzo: 
                  type: number
                  description: 'Price of the dish'
                descrizione: 
                  type: string
                  description: 'Brief description of the dish (quantities, ingredients, etc)'
                foto: 
                  type: string
                  description: 'Path to the photo (may be local or online)'
                managerpwd:
                  type: string
                  description: 'Input from the manager needed to authenticate the manager'
            example:
              nome: 'Nigiri di Tonno'  
              prezzo: 3.99
              descrizione: '8 pezzi. Ingredienti: riso, tonno crudo'
              foto: 'www.fotoDelPiatto.com'
              managerpwd: 'admin'
      responses:
        '201':
          description: 'Dish created. Link in the Location header'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Piatto'
              example:
                nome: 'Nigiri di Tonno'  
                prezzo: 3.99
                descrizione: '8 pezzi. Ingredienti: riso, tonno crudo'
                foto: 'www.fotoDelPiatto.com'
        '400':
          description: 'Name was submitted empty'
        '400':
          description: 'Price was submitted empty'
        '400':
          description: 'Price was NaN'
        '403':
          description: 'Manager authentication failed'
        '404':
          description: 'Restaurant not found'
  /piattosRisto/eliminaPiatto/{id}/{managerpwd}:
    delete:
      description: >-
          Delete a dish from the menu of a specific restaurant. 
      summary: Delete a dish
      parameters:
        - name: x-access-token
          in: header
          description: an authorization header
          required: true
          schema:
            type: string 
          example: token.esempio.prova
        - name: id
          in: path
          required: true
          description: the ID of the dish to delete
          schema:
            type: string
          example: 'bagwr253678er9oidjhu7'
        - name: managerpwd
          in: path
          required: true
          description: the password of the manager
          schema:
            type: string
          example: 'admin'
      responses:
        '204':
          description: 'No content'
        '403':
          description: 'Manager authentication failed'
        '404':
          description: 'Error: could not find the restaurant'
        '404':
          description: 'Error: could not find the dish you are trying to eliminate'
  /piattosRisto/ordineTavolo/{idTavolo}:
    get:
      description: >-
          Chiamata API in GET che dato nei parametri l'id di un tavolo restituisce la lista dei piatti nell'ordine del tavolo
      summary: Mostra ordine tavolo
      parameters:
        - name: idTavolo
          in: path
          description: Id del tavolo
          required: true
          schema:
            type: string
          example: '629a4cb56815e55554e7d7ae'
        - name: x-access-token
          in: header
          description: an authorization header
          required: true
          schema:
            type: string
          example: token.esempio.prova
      responses:
        '201':
          description: 'Ordine del tavolo'
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      description: 'Id del piatto'
                    nome:
                      type: string
                      description: 'Nome del piatto'
                    foto:
                      type: string
                      description: 'link alla foto del piatto'
                    stato:
                      type: string
                      description: 'Stato del tavolo'
              example:
                  - id: '629a4cb56815e55554e7d7ae'
                    nome: 'Nigiti al salmone'
                    foto: 'www.foto.it'
                    stato: 'in consegna'
                  - id: '629715ffe86e996e2f47ff5c'
                    nome: 'Nigiti al tonno'
                    foto: 'www.foto.it'
                    stato: 'consegnato'
        '404':
            description: 'Tavolo non trovato'
        '404':
            description: 'Ordine non trovato'
  /piattosRisto/cambiaStato:
    post:
      description: >-
          Chiamata API in POST che dato nel body l'id del tavolo e del piatto e lo stato cambia lo stato del piatto in quello desiderato
      summary: Cambia stato di un piatto
      requestBody:
        content:
            application/json:
              schema:
                type: object
                required:
                  - idT
                  - idP
                  - stato
                properties:
                  idT:
                    type: string
                    description: 'Id del Tavolo'
                  idP:
                    type: number
                    description: 'Id del piatto'
                  stato:
                    type: string
                    description: 'Nuovo stato del piatto'
              example:
                  idT: '629a4cb56815e55554e7d7ae'
                  idP: '629715ffe86e996e2f47ff5c'
                  stato: 'consegnato'
      responses:
        '201':
          description: 'Stato cambiato al piatto'
        '405':
            description: 'Tavolo non trovato'
...
components:
  schemas:
    Piatto:
      type: object
      required: 
      - nome
      - prezzo
      properties: 
        nome: 
          type: string
          description: 'Name of the dish'
        prezzo: 
          type: number
          description: 'Price of the dish'
        descrizione: 
          type: string
          description: 'Description of the dish'
        foto: 
          type: string
          description: 'path to the photo of the dish'
        stato: 
          type: string
          description: 'status of the dish (in preparazione/consegnato/in consegna)'
      description: 'dishes in the database'
...